Move a few implementation details of IPConverter into the cpp file.
